refactor(cores): fix the import path
fix the import path in cores cause of renaming core to cores
