+++ tsconfig.json
@@ -16,7 +16,7 @@
     "incremental": true,
     "types": ["@testing-library/jest-dom"],
     "paths": {
-      "@/core/*": ["./src/core/*"],
+      "@/cores/*": ["./src/cores/*"],
       "@/hooks/*": ["./src/hooks/*"],
       "@/libs/*": ["./src/libs/*"],
       "@/utils/*": ["./src/utils/*"],
